requirements.txt

celery==5.5.3
Django==5.2.3
django-environ==0.12.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
python-telegram-bot==22.1
redis==6.2.0

